Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/character/skeleton_state.cpp
@@ -330,7 +330,7 @@ localTransformToJointParametersRespectingConstraints(
 
   // Determine which rotation axes are active
 
-  std::array<size_t, 3> activeAxes{};
+  std::array<int, 3> activeAxes{};
   size_t numActiveAxes = 0;
   for (int k = 0; k < 3; ++k) {
     if (activeJointParams[paramOffset + 3 + k]) {
